fix: register test case
