+++ packages/xhear/tests/statics/register-test.html
@@ -19,8 +19,8 @@
     <test-ele></test-ele>
     
     <script type="module">
-      // import $ from "../../src/base.mjs";
-      // window.$ = $;
+      import $ from "../../src/base.mjs";
+      window.$ = $;
 
       // $("test-ele").aaa = "aaa";
